requirements.txt
langchain
langchain-community
langchain-core
termcolor
python-dotenv
pymongo
certifi
sentence-transformers
fastapi
uvicorn
faiss-cpu 